Organize java files into folders/packages
