--- app/src/main/java/com/example/trialio/Barcode.java
@@ -1,13 +1,0 @@
-package com.example.trialio;
-
-public class Barcode {
-
-    private Experiment experiment;
-    private Trial trial;
-    private String barcode; // change type later
-
-    public void setBarcodeImage(String barcode) {
-
-    }
-
-}
