--- app/src/main/java/com/example/trialio/BarcodeManager.java
@@ -1,22 +1,0 @@
-package com.example.trialio;
-
-import java.util.Collection;
-
-public class BarcodeManager {
-
-    public Collection<Barcode> barcodes;
-
-    // public void generateQRCode (Experiment experiment, Trial trial) { }
-
-    // public void registerBarcode (Experiment experiment, Trial trial, Image barcodeImage) { }
-
-    // public Collection<Barcode> getBarcodes() { }
-
-    // public void removeBarcode (Barcode barcode) { }
-
-
-
-
-
-
-}
