--- app/src/main/java/com/example/trialio/CountTrial.java
@@ -1,21 +1,0 @@
-package com.example.trialio;
-
-import java.io.Serializable;
-import java.util.Date;
-
-public class CountTrial extends Trial implements Serializable {
-    private int count;
-
-    public CountTrial(String experimenterID, com.example.trialio.Location location, Date date) {
-        super(experimenterID, location, date);
-        this.count = 1;
-    }
-
-    public int getCount() {
-        return count;
-    }
-
-    public void setCount(int count) {
-        this.count = count;
-    }
-}
