--- app/src/main/java/com/example/trialio/Post.java
@@ -1,12 +1,0 @@
-package com.example.trialio;
-
-public abstract class Post {
-
-    private String body;
-    private User user;
-
-    public String getBody() {
-        return this.body;
-    }
-
-}
