--- app/src/main/java/com/example/trialio/Question.java
@@ -1,14 +1,0 @@
-package com.example.trialio;
-
-import java.io.Serializable;
-import java.util.Collection;
-
-public class Question implements Serializable {
-
-    private String title;
-    private Collection<Reply> replies;
-
-//    public Collection<Reply> getAllReplies () {    }
-
-
-}
