--- app/src/main/java/com/example/trialio/Reply.java
@@ -1,12 +1,0 @@
-package com.example.trialio;
-
-import java.io.Serializable;
-import java.util.Collection;
-
-public class Reply implements Serializable {
-
-    private Collection<Reply> replies;
-
-//    public Collection<Reply> getAllReplies() { }
-
-}
