--- app/src/main/java/com/example/trialio/StatisticsUtility.java
@@ -1,20 +1,0 @@
-package com.example.trialio;
-
-import java.util.ArrayList;
-
-public class StatisticsUtility {
-
-    public ArrayList<Double> getExperimentStatistics (String type, Experiment experiment) {
-        //...
-        return new ArrayList<Double>();
-    }
-
-    public void getTrialsHistogram (String type, Experiment experiment) {
-        //...
-    }
-
-    public void getTrialsPlot (String type, Experiment experiment) {
-        //...
-    }
-
-}
