+++ app/src/main/java/com/example/trialio/controllers/BarcodeManager.java
@@ -1,0 +1,24 @@
+package com.example.trialio.controllers;
+
+import com.example.trialio.models.Barcode;
+
+import java.util.Collection;
+
+public class BarcodeManager {
+
+    public Collection<Barcode> barcodes;
+
+    // public void generateQRCode (Experiment experiment, Trial trial) { }
+
+    // public void registerBarcode (Experiment experiment, Trial trial, Image barcodeImage) { }
+
+    // public Collection<Barcode> getBarcodes() { }
+
+    // public void removeBarcode (Barcode barcode) { }
+
+
+
+
+
+
+}
