+++ app/src/main/java/com/example/trialio/models/Barcode.java
@@ -1,0 +1,13 @@
+package com.example.trialio.models;
+
+public class Barcode {
+
+    private Experiment experiment;
+    private Trial trial;
+    private String barcode; // change type later
+
+    public void setBarcodeImage(String barcode) {
+
+    }
+
+}
