+++ app/src/main/java/com/example/trialio/models/BinomialTrial.java
@@ -1,0 +1,21 @@
+package com.example.trialio.models;
+
+import java.io.Serializable;
+import java.util.Date;
+
+public class BinomialTrial extends Trial implements Serializable {
+    private boolean isSuccess;
+
+    public BinomialTrial(String experimenterID, Location location, Date date, boolean isSuccess) {
+        super(experimenterID, location, date);
+        this.isSuccess = isSuccess;
+    }
+
+    public boolean getIsSuccess() {
+        return isSuccess;
+    }
+
+    public void setIsSuccess(boolean success) {
+        isSuccess = success;
+    }
+}
