+++ app/src/main/java/com/example/trialio/models/Location.java
@@ -1,0 +1,31 @@
+package com.example.trialio.models;
+
+public class Location {
+    private double latitude;
+    private double longitude;
+
+    public Location() { }
+
+    public Location(double latitude, double longitude) {
+        this.latitude = latitude;
+        this.longitude = longitude;
+    }
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(double latitude) {
+        this.latitude = latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(double longitude) {
+        this.longitude = longitude;
+    }
+
+    //    public ArrayList<double> getCoord() { }
+}
