+++ app/src/main/java/com/example/trialio/models/Question.java
@@ -1,0 +1,14 @@
+package com.example.trialio.models;
+
+import java.io.Serializable;
+import java.util.Collection;
+
+public class Question implements Serializable {
+
+    private String title;
+    private Collection<Reply> replies;
+
+//    public Collection<Reply> getAllReplies () {    }
+
+
+}
