+++ app/src/main/java/com/example/trialio/models/Reply.java
@@ -1,0 +1,12 @@
+package com.example.trialio.models;
+
+import java.io.Serializable;
+import java.util.Collection;
+
+public class Reply implements Serializable {
+
+    private Collection<Reply> replies;
+
+//    public Collection<Reply> getAllReplies() { }
+
+}
