+++ app/src/main/java/com/example/trialio/utils/StatisticsUtility.java
@@ -1,0 +1,22 @@
+package com.example.trialio.utils;
+
+import com.example.trialio.models.Experiment;
+
+import java.util.ArrayList;
+
+public class StatisticsUtility {
+
+    public ArrayList<Double> getExperimentStatistics (String type, Experiment experiment) {
+        //...
+        return new ArrayList<Double>();
+    }
+
+    public void getTrialsHistogram (String type, Experiment experiment) {
+        //...
+    }
+
+    public void getTrialsPlot (String type, Experiment experiment) {
+        //...
+    }
+
+}
