+++ app/src/main/res/layout/activity_experiment.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#535353"
-    tools:context=".ExperimentActivity">
+    tools:context=".activities.ExperimentActivity">
 
     <LinearLayout
         android:id="@+id/linearLayout"
